+++ language/source/codegen/abstract_syntax_tree/keywords/types/unsigned_integers/keyword_u16_node.h
@@ -7,12 +7,12 @@ namespace channel {
 	 */
 	class keyword_u16_node : public node {
 	public:
-		keyword_u16_node(u16 value);
+		keyword_u16_node(u64 value);
 		value* accept(visitor& visitor) override;
 		std::string get_node_name() const override;
 
-		u16 get_value() const;
+		u64 get_value() const;
 	private:
-		u16 m_value;
+		u64 m_value;
 	};
 }
